% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfcop.R
\name{dfcop}
\alias{dfcop}
\title{Discrete factor copula models}
\usage{
dfcop(data, family_set = "all", selcrit = "bic", ngrid = 100,
  keep_data = TRUE)
}
\arguments{
\item{data}{a binary matrix or data.frame.}

\item{family_set}{a character vector of families; as in \code{dfcop_dist()},
see \emph{Details} for additional options.}

\item{selcrit}{criterion for family selection, either \code{"loglik"}, \code{"aic"}, or
\code{"bic"}.}

\item{ngrid}{number of nodes and weights for the Gaussian quadrature.}

\item{keep_data}{whether the data should be stored (necessary for computing
fit statistics and using \code{fitted()}).}
}
\value{
An object inherting from \code{dfcop} and \code{dfcop_dist}.
}
\description{
Discrete factor copula models
}
\details{
In addition, to the families in \code{dfcop_dist()}, the following convenience
definition from  can be used (and combined):
\describe{
\item{"all"}{all families}.
\item{"archimedean"}{archimedean families.}
\item{"elliptical"}{elliptical families.}
\item{"bbs"}{BB families.}
\item{"oneparametric "}{one parameter families.}
\item{"twoparametric "}{two parameter families.}
}
Partial matching is activated. For example, you can write \code{"arch"} instead
of the full name \code{"archimedean"}.
}
\examples{
# Sample size, dimension, correlation parameter
n <- 1e3
d <- 1e1
rho <- 0.7

# The model
prob <- runif(d)
dfcop <- dfcop_dist(prob, "gaussian", rho)

# Data
X <- rdfcop(n, dfcop)

# Fit
fit <- dfcop(X)
fit$dist

}
