% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfcop_dist.R
\name{dfcop_dist}
\alias{dfcop_dist}
\alias{ddfcop}
\alias{rdfcop}
\title{A discrete factor copula distribution}
\usage{
dfcop_dist(prob, family = "indep", parameters = numeric(0))

ddfcop(x, prob, family = "indep", parameters = numeric(0), ngrid = 100)

rdfcop(n, prob, family = "indep", parameters = numeric(0))
}
\arguments{
\item{prob}{the marginal probabilities, a vector of numbers in (0,1).}

\item{family}{the copula family, a string containing the family name (see
\emph{Details} for all possible families).}

\item{parameters}{a vector or matrix of copula paramters.}

\item{x}{a binary vector of the same length as prob}

\item{ngrid}{number of nodes and weights for the Gaussian quadrature}

\item{n}{number of observations.}
}
\value{
An object of class \code{dfcop_dist}.
}
\description{
A S3 class to store discrete factor copula distributions.
}
\details{
The implemented families listed below. Partial matching is activated, i.e.,
\code{"gauss"} is equivalent to \code{"gaussian"}.
\describe{
\item{\code{indep}}{Independence copula.}
\item{\code{gaussian}}{Gaussian copula.}
\item{\code{t}}{Student t copula.}
\item{\code{clayton}}{Clayton copula.}
\item{\code{gumbel}}{Gumbel copula.}
\item{\code{frank}}{Frank copula.}
\item{\code{joe}}{Joe copula.}
\item{\code{bb1}}{BB1 copula.}
\item{\code{bb6}}{BB6 copula.}
\item{\code{bb7}}{BB7 copula.}
\item{\code{bb8}}{BB8 copula.}
}
}
\note{
The evaluation functions can optionally be used with a \code{dfcop_dist} object,
e.g., \code{ddfcop(c(1, 0), dfcop_dist(c(0.5, 0.5),"indep"))}.
}
\examples{
# A 10-dimensional Gaussian factor copula
dfcop_dist(runif(10),"gaussian", 0.5)
str(dfcop_dist(runif(10), "gauss", 0.5))

# A 100-dimensional rotated Clayton factor copula
dfcop <- dfcop_dist(runif(100), "clayton", 3)
 
# evaluate the probability mass function
ddfcop(cbind(c(1,1,0,0), c(1,0,1,0)), c(0.5, 0.5), "indep")
ddfcop(rbinom(100, 1, 0.5), dfcop)

# simulate data
rdfcop(1e2, c(0.5, 0.5))
rdfcop(1e2, dfcop)
}
