% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfcop.R
\name{predict.dfcop}
\alias{predict.dfcop}
\alias{fitted.dfcop}
\title{Predictions and fitted values for a discrete factor copula model}
\usage{
\method{predict}{dfcop}(object, newdata, ...)

\method{fitted}{dfcop}(object, ...)
}
\arguments{
\item{object}{a \code{dfcop} object.}

\item{newdata}{points where the fit shall be evaluated.}

\item{...}{unused.}
}
\description{
Predictions and fitted values for a discrete factor copula model
}
\examples{
# Sample size, dimension, correlation parameter
n <- 1e3
d <- 1e1
rho <- 0.7

# The model
prob <- runif(d)
dfcop <- dfcop_dist(prob, "gaussian", rho)

# Data
X <- rdfcop(n, dfcop)

# Fit
fit <- dfcop(X)
all.equal(predict(fit, X), fitted(fit))

}
