% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfcop.R
\name{predict.dfcop}
\alias{predict.dfcop}
\alias{fitted.dfcop}
\alias{logLik.dfcop}
\alias{AIC.dfcop}
\alias{BIC.dfcop}
\alias{fitted.dfcop}
\alias{logLik.dfcop}
\title{S3 methods for discrete factor copula model}
\usage{
\method{predict}{dfcop}(object, newdata, ...)

\method{fitted}{dfcop}(object, ...)

\method{logLik}{dfcop}(object, ...)
}
\arguments{
\item{object}{a \code{dfcop} object.}

\item{newdata}{points where the fit shall be evaluated.}

\item{...}{unused.}
}
\description{
Currently available:
\describe{
\item{\code{predict}}{Predict the probabilities of new observations.}
\item{\code{fitted}}{Fitted probabilities.}
\item{\code{logLik}}{Log-likelihood at the MLE.}
\item{\code{AIC}}{Akaike's Information Criterion (see \link{AIC}).}
\item{\code{BIC}}{Bayesian Information Criterion (see \link{BIC}).}
}
}
\examples{
# Sample size, dimension, correlation parameter
n <- 1e3
d <- 1e1
rho <- 0.7

# The model
prob <- runif(d)
dfcop <- dfcop_dist(prob, "gaussian", rho)

# Data
X <- rdfcop(n, dfcop)

# Fit
fit <- dfcop(X)
all.equal(predict(fit, X), fitted(fit))

# Other methods
logLik(fit)
AIC(fit)
BIC(fit)

}
